requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pillow>=10.0.0
matplotlib>=3.8.0
pandas>=2.1.0
jupyter
ipykernel